After review: move commands to internal
