+++ internal/cli/commands/delete_creds.go
