+++ internal/cli/commands/logout.go
