+++ internal/cli/commands/register.go
